Fix bug that causes conversion error "ClassCastException: java.lang.Double cannot be cast to java.lang.Float"

+++ portfolio/src/main/java/com/google/sps/servlets/DataServlet.java
@@ -76,7 +76,7 @@ public class DataServlet extends HttpServlet {
               (long) entity.getKey().getId(),
               (long) entity.getProperty("likeCount"),
               (String) entity.getProperty("email"),
-              (float) entity.getProperty("sentimentScore"))
+              (float)(double) entity.getProperty("sentimentScore"))
           )
       .collect(Collectors.toList());
